refactor(storage): restructure to top-level, fix file_ops web recording, clean container
- Move core/storage/ → storage/ (top-level package, proper separation from core)
- Fix file_operations=0 bug in web mode: streaming_service never set current_checkpoint_id,
  so _record_operation silently skipped all writes; fix by adding current_run_id ContextVar
  to sandbox/thread_context and calling set_current_run_id(run_id) in _run_agent_to_buffer
- Decouple core/filesystem/middleware.py from tui: replace tui.operations.current_checkpoint_id
  import with sandbox.thread_context.get_current_run_id()
- Remove dead supabase_bindings param from StorageContainer (no caller ever used it)
- Collapse 6×2 duplicate _build_supabase_* methods into generic _build_repo + registry dict
- Add vite.supabase.config.ts for running frontend against supabase-mode backend (port 18001)
- Fix stale _get_conn monkeypatch in test_storage_runtime_wiring.py (attr was already removed)
