+++ backend/web/services/agent_pool.py
@@ -8,7 +8,7 @@ from typing import Any
 from fastapi import FastAPI
 
 from agent import create_leon_agent
-from core.storage.runtime import build_storage_container
+from storage.runtime import build_storage_container
 from sandbox.manager import lookup_sandbox_for_thread
 from sandbox.thread_context import set_current_thread_id
