+++ frontend/app/vite.supabase.config.ts
@@ -1,0 +1,33 @@
+import path from 'path'
+import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite'
+import { inspectAttr } from 'kimi-plugin-inspect-react'
+
+export default defineConfig({
+  base: './',
+  plugins: [inspectAttr(), react()],
+  server: {
+    host: '0.0.0.0',
+    port: 5180,
+    allowedHosts: true,
+    proxy: {
+      '/api': {
+        target: 'http://127.0.0.1:18001',
+        changeOrigin: true,
+        configure: (proxy) => {
+          proxy.on('proxyRes', (proxyRes) => {
+            if (proxyRes.headers['content-type']?.includes('text/event-stream')) {
+              proxyRes.headers['cache-control'] = 'no-cache'
+              proxyRes.headers['x-accel-buffering'] = 'no'
+            }
+          })
+        },
+      },
+    },
+  },
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
+    },
+  },
+})
