+++ storage/container.py
@@ -1,0 +1,161 @@
+"""Storage container with repo-level provider selection."""
+
+from __future__ import annotations
+
+import importlib
+from collections.abc import Mapping
+from pathlib import Path
+from typing import Any, Literal
+import sqlite3
+
+from .contracts import (
+    CheckpointRepo,
+    EvalRepo,
+    FileOperationRepo,
+    RunEventRepo,
+    SummaryRepo,
+    ThreadConfigRepo,
+)
+
+StorageStrategy = Literal["sqlite", "supabase"]
+RepoProviderMap = Mapping[str, str]
+
+# @@@repo-registry - maps repo name → (supabase module path, class name)
+_REPO_REGISTRY: dict[str, tuple[str, str]] = {
+    "checkpoint_repo":     ("storage.providers.supabase.checkpoint_repo",      "SupabaseCheckpointRepo"),
+    "thread_config_repo":  ("storage.providers.supabase.thread_config_repo",   "SupabaseThreadConfigRepo"),
+    "run_event_repo":      ("storage.providers.supabase.run_event_repo",       "SupabaseRunEventRepo"),
+    "file_operation_repo": ("storage.providers.supabase.file_operation_repo",  "SupabaseFileOperationRepo"),
+    "summary_repo":        ("storage.providers.supabase.summary_repo",         "SupabaseSummaryRepo"),
+    "eval_repo":           ("storage.providers.supabase.eval_repo",            "SupabaseEvalRepo"),
+}
+
+
+class StorageContainer:
+    """Composition root for storage repos."""
+
+    _SUPPORTED_STRATEGIES = {"sqlite", "supabase"}
+    _REPO_NAMES = tuple(_REPO_REGISTRY.keys())
+
+    def __init__(
+        self,
+        main_db_path: str | Path | None = None,
+        eval_db_path: str | Path | None = None,
+        strategy: StorageStrategy = "sqlite",
+        repo_providers: RepoProviderMap | None = None,
+        supabase_client: Any | None = None,
+    ) -> None:
+        if strategy not in self._SUPPORTED_STRATEGIES:
+            raise ValueError(
+                f"Unsupported storage strategy: {strategy}. "
+                f"Supported strategies: {', '.join(sorted(self._SUPPORTED_STRATEGIES))}"
+            )
+        root = Path.home() / ".leon"
+        self._main_db = Path(main_db_path) if main_db_path else root / "leon.db"
+        self._eval_db = Path(eval_db_path) if eval_db_path else root / "eval.db"
+        self._strategy: StorageStrategy = strategy
+        self._supabase_client = supabase_client
+        self._repo_providers = self._resolve_repo_providers(
+            default_strategy=strategy,
+            repo_providers=repo_providers,
+        )
+
+    def checkpoint_repo(self) -> CheckpointRepo:
+        return self._build_repo("checkpoint_repo", self._sqlite_checkpoint_repo)
+
+    def thread_config_repo(self) -> ThreadConfigRepo:
+        return self._build_repo("thread_config_repo", self._sqlite_thread_config_repo)
+
+    def run_event_repo(self) -> RunEventRepo:
+        return self._build_repo("run_event_repo", self._sqlite_run_event_repo)
+
+    def file_operation_repo(self) -> FileOperationRepo:
+        return self._build_repo("file_operation_repo", self._sqlite_file_operation_repo)
+
+    def summary_repo(self) -> SummaryRepo:
+        return self._build_repo("summary_repo", self._sqlite_summary_repo)
+
+    def eval_repo(self) -> EvalRepo:
+        return self._build_repo("eval_repo", self._sqlite_eval_repo)
+
+    def provider_for(self, repo_name: str) -> StorageStrategy:
+        return self._provider_for(repo_name)
+
+    # --- internal ---
+
+    def _build_repo(self, name: str, sqlite_factory):
+        if self._provider_for(name) == "supabase":
+            if self._supabase_client is None:
+                raise RuntimeError(
+                    f"Supabase strategy {name} requires supabase_client. "
+                    "Pass supabase_client=... into StorageContainer."
+                )
+            mod_path, cls_name = _REPO_REGISTRY[name]
+            mod = importlib.import_module(mod_path)
+            return getattr(mod, cls_name)(client=self._supabase_client)
+        return sqlite_factory()
+
+    def _provider_for(self, repo_name: str) -> StorageStrategy:
+        if repo_name not in self._REPO_NAMES:
+            supported = ", ".join(self._REPO_NAMES)
+            raise ValueError(f"Unknown repo name: {repo_name}. Supported repo names: {supported}")
+        return self._repo_providers[repo_name]
+
+    @classmethod
+    def _resolve_repo_providers(
+        cls,
+        *,
+        default_strategy: StorageStrategy,
+        repo_providers: RepoProviderMap | None,
+    ) -> dict[str, StorageStrategy]:
+        overrides: Mapping[str, Any] = repo_providers or {}
+        unknown_repos = sorted(set(overrides.keys()) - set(cls._REPO_NAMES))
+        if unknown_repos:
+            supported = ", ".join(cls._REPO_NAMES)
+            unknown = ", ".join(unknown_repos)
+            raise ValueError(f"Unknown repo provider bindings: {unknown}. Supported repo names: {supported}")
+
+        resolved: dict[str, StorageStrategy] = {name: default_strategy for name in cls._REPO_NAMES}
+        # @@@repo-provider-override - default strategy keeps current behavior; only explicitly listed repos diverge.
+        for repo_name, provider in overrides.items():
+            if not isinstance(provider, str):
+                raise ValueError(
+                    f"Invalid provider value for {repo_name}: {provider!r}. Expected 'sqlite' or 'supabase'."
+                )
+            normalized = provider.strip().lower()
+            if normalized not in cls._SUPPORTED_STRATEGIES:
+                supported = ", ".join(sorted(cls._SUPPORTED_STRATEGIES))
+                raise ValueError(
+                    f"Unsupported provider for {repo_name}: {provider!r}. Supported providers: {supported}"
+                )
+            resolved[repo_name] = normalized
+        return resolved
+
+    # SQLite factories (called only when provider == "sqlite")
+
+    def _sqlite_checkpoint_repo(self) -> CheckpointRepo:
+        from storage.providers.sqlite.checkpoint_repo import SQLiteCheckpointRepo
+        return SQLiteCheckpointRepo(db_path=self._main_db)
+
+    def _sqlite_thread_config_repo(self) -> ThreadConfigRepo:
+        from storage.providers.sqlite.thread_config_repo import SQLiteThreadConfigRepo
+        return SQLiteThreadConfigRepo(db_path=self._main_db)
+
+    def _sqlite_run_event_repo(self) -> RunEventRepo:
+        from storage.providers.sqlite.run_event_repo import SQLiteRunEventRepo
+        return SQLiteRunEventRepo(db_path=self._main_db)
+
+    def _sqlite_file_operation_repo(self) -> FileOperationRepo:
+        from storage.providers.sqlite.file_operation_repo import SQLiteFileOperationRepo
+        return SQLiteFileOperationRepo(db_path=self._main_db)
+
+    def _sqlite_summary_repo(self) -> SummaryRepo:
+        from storage.providers.sqlite.summary_repo import SQLiteSummaryRepo
+        return SQLiteSummaryRepo(
+            db_path=self._main_db,
+            connect_fn=lambda p: sqlite3.connect(str(p)),
+        )
+
+    def _sqlite_eval_repo(self) -> EvalRepo:
+        from storage.providers.sqlite.eval_repo import SQLiteEvalRepo
+        return SQLiteEvalRepo(db_path=self._eval_db)
