+++ storage/providers/supabase/file_operation_repo.py
@@ -7,7 +7,7 @@ import time
 import uuid
 from typing import Any
 
-from core.storage.providers.sqlite.file_operation_repo import FileOperationRow
+from storage.providers.sqlite.file_operation_repo import FileOperationRow
 
 
 class SupabaseFileOperationRepo:
