+++ tests/test_run_event_repo.py
@@ -2,8 +2,8 @@ import sqlite3
 
 import pytest
 
-from core.storage.providers.sqlite.run_event_repo import SQLiteRunEventRepo
-from core.storage.providers.supabase.run_event_repo import SupabaseRunEventRepo
+from storage.providers.sqlite.run_event_repo import SQLiteRunEventRepo
+from storage.providers.supabase.run_event_repo import SupabaseRunEventRepo
 
 
 def test_append_and_list_events_with_cursor(tmp_path):
