+++ tests/test_storage_runtime_wiring.py
@@ -12,9 +12,9 @@ import pytest
 from backend.web.services import agent_pool
 from backend.web.services.event_buffer import RunEventBuffer
 from backend.web.services.streaming_service import _run_agent_to_buffer
-from core.storage.providers.sqlite.checkpoint_repo import SQLiteCheckpointRepo
-from core.storage.providers.sqlite.eval_repo import SQLiteEvalRepo
-from core.storage.providers.supabase.checkpoint_repo import SupabaseCheckpointRepo
+from storage.providers.sqlite.checkpoint_repo import SQLiteCheckpointRepo
+from storage.providers.sqlite.eval_repo import SQLiteEvalRepo
+from storage.providers.supabase.checkpoint_repo import SupabaseCheckpointRepo
 
 
 class _FakeSupabaseClient:
@@ -250,13 +250,6 @@ class _FakeRuntimeAgent:
 
 def test_run_runtime_consumes_storage_container_run_event_repo(monkeypatch: pytest.MonkeyPatch) -> None:
     async def _run() -> None:
-        import backend.web.services.event_store as event_store
-
-        async def _raise_if_sqlite_conn_used() -> Any:
-            raise AssertionError("sqlite connection path should not be used when storage container repo is present")
-
-        monkeypatch.setattr(event_store, "_get_conn", _raise_if_sqlite_conn_used)
-
         repo = _FakeRunEventRepo()
         agent = _FakeRuntimeAgent(storage_container=_FakeStorageContainer(repo))
         app = SimpleNamespace(state=SimpleNamespace(thread_tasks={}, thread_event_buffers={}))
